Password protected board bug find

+++ client/src/main/java/client/scenes/BoardOverviewCtrl.java
@@ -122,13 +122,11 @@ public class BoardOverviewCtrl implements Initializable {
             try {
                 File current = new File(boardFilePath);
                 Scanner scanner = new Scanner(current);
-                // Vbox to contain all boards.
                 VBox boardsBox = new VBox(25);
                 int size = 0;
                 while (scanner.hasNextLine()) {
                     String line = scanner.nextLine();
                     long boardID = Long.parseLong(line.split(" -BOOL- ")[0]);
-
                     boolean created = line.split(" -BOOL- ")[1].equals("true");
                     try {
                         Board b = server.getBoardById(boardID);
@@ -139,18 +137,14 @@ public class BoardOverviewCtrl implements Initializable {
                         ImageView imageView = new ImageView(new Image(Objects.requireNonNull(getClass().getResourceAsStream("/client/assets/key.png"))));
                         Button openBoard = new Button("Open");
                         Button delete = new Button("X");
-
                         String lock = "unlock";
-                        if (b.isLocked()) {
-                            lock = "lock";
-                        }
+                        if (b.isLocked()) lock = "lock";
                         ImageView locked = new ImageView(new Image(Objects.requireNonNull(getClass().getResourceAsStream("/client/assets/" + lock + ".png"))));
                         locked.setPreserveRatio(true);
                         locked.setPickOnBounds(true);
                         locked.setFitHeight(40);
 
                         prepareContent(boardLabel, copyKey, imageView, openBoard, delete, b, created);
-
                         boardContent.getChildren().addAll(boardLabel, copyKey, locked, openBoard, delete);
                         boardsBox.getChildren().add(boardContent);
                     } catch (BadRequestException e) {
@@ -159,9 +153,7 @@ public class BoardOverviewCtrl implements Initializable {
                 }
                 scanner.close();
                 if (!current.getName().equals(boardFilePath)) {
-                    if (!current.getName().equals("boardsTemp.txt")) {
-                        new File("boardsTemp.txt").delete();
-                    }
+                    if (!current.getName().equals("boardsTemp.txt")) new File("boardsTemp.txt").delete();
                     new File(boardFilePath).delete();
                     current.renameTo(new File(boardFilePath));
                     new File("boardsTest.txt").delete();
@@ -461,16 +453,25 @@ public class BoardOverviewCtrl implements Initializable {
             Label boardLabel = new Label(board.getName());
             Button copyKey = new Button();
             ImageView imageView = new ImageView(new Image(Objects.requireNonNull(getClass().getResourceAsStream("/client/assets/key.png"))));
-            Button openBoard = new Button("Open Board");
+            Button openBoard = new Button("Open");
             Button delete = new Button("X");
+
+            String lock = "unlock";
+            if (board.isLocked()) {
+                lock = "lock";
+            }
+            ImageView locked = new ImageView(new Image(Objects.requireNonNull(getClass().getResourceAsStream("/client/assets/" + lock + ".png"))));
+            locked.setPreserveRatio(true);
+            locked.setPickOnBounds(true);
+            locked.setFitHeight(40);
 
             prepareContent(boardLabel, copyKey, imageView, openBoard, delete, board, true);
 
-            boardContent.getChildren().addAll(boardLabel, copyKey, openBoard, delete);
+            boardContent.getChildren().addAll(boardLabel, copyKey, locked, openBoard, delete);
             boardsBox.getChildren().add(boardContent);
-            boardsBox.setPrefSize(600, 225 * size);
-            boardContainer.setContent(boardsBox);
         }
+        boardsBox.setPrefSize(600, 225 * size);
+        boardContainer.setContent(boardsBox);
     }
 
     /**
